+++ python_enigma/__init__.py
@@ -1,3 +1,3 @@
 __version__ = "1.1.4"
 
-from .enigma import *
+# from .enigma import *