+++ python_enigma/resources/__init__.py
@@ -1,0 +1,3 @@
+# Make this resource directory a package for importlib_resources
+
+__version__ = "1.0.0"